requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
arrow==1.2.2
async-timeout==4.0.2
attrs==21.4.0
babelfish==0.6.0
beautifulsoup4==4.11.1
cattrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cinemagoer==2022.12.4
commonmark==0.9.1
decorating==0.6.1
docopt==0.6.2
easyocr==1.5.0
exceptiongroup==1.0.0rc8
frozenlist==1.3.1
greenlet==2.0.1
guessit==3.4.3
idna==3.3
imageio==2.19.5
importlib-resources==5.9.0
iniconfig==1.1.1
inquirerpy==0.3.4
jellyfish==0.9.0
jikanpy==4.3.2
langcodes==3.3.0
language-data==1.1
lxml==4.9.1
marisa-trie==0.7.7
multidict==6.0.2
natsort==8.1.0
networkx==2.8.5
num2words==0.5.10
numpy==1.23.1
opencv-python-headless==4.5.4.60
orjson==3.7.12
packaging==21.3
pfzy==0.3.4
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.30
py==1.11.0
pycdlib==1.13.0
pycparser==2.21
Pygments==2.13.0
pymediainfo==5.1.0
pymplschapters==1.0.4
pynumparser==1.4.1
pyparsing==3.0.9
pytest==7.1.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
PyYAML==6.0
rebulk==3.1.0
requests==2.28.1
requests-cache==0.9.5
rich==12.6.0
scikit-image==0.19.3
scipy==1.8.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
srt==3.5.2
tesserocr==2.5.2
tifffile==2022.5.4
tmdbsimple==2.9.1
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.1
typing-extensions==4.3.0
url-normalize==1.4.3
urllib3==1.26.11
vosk==0.3.44
wcwidth==0.2.5
websockets==10.4
xxhash==3.1.0
yarl==1.8.1
zipp==3.8.1
